add healing

+++ module03/ex00/ClapTrap.hpp
@@ -27,9 +27,9 @@ class ClapTrap {
         void takeDamage(unsigned int amount);
         void beRepaired(unsigned int amount);  
         std::string getName() const;
-        uint8_t     getAttack() const;
-        uint8_t     getEnergy() const;
-        uint8_t     getHp() const;
+        uint32_t     getAttack() const;
+        uint32_t     getEnergy() const;
+        int32_t     getHp() const;
         void        setAttackDamage(uint8_t amount);
         ClapTrap(std::string name);        
     private:
